lv3 토큰으로 게시물 접근 구현

+++ src/main/java/com/example/bloglv1/service/PostService.java
@@ -5,12 +5,14 @@ import com.example.bloglv1.dto.PostResponseDto;
 import com.example.bloglv1.entity.Post;
 import com.example.bloglv1.entity.User;
 import com.example.bloglv1.repository.PostRepository;
+import com.example.bloglv1.security.UserDetailsImpl;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.RejectedExecutionException;
 
 @Service
 @RequiredArgsConstructor
@@ -19,13 +21,14 @@ public class PostService {
     private final PostRepository postRepository;
 
     //게시물 작성
-    public PostResponseDto createPost(PostRequestDto postRequestDto) {
+    public PostResponseDto createPost(User user, PostRequestDto postRequestDto) {
 
         //게시글 생성
-        Post post = new Post(postRequestDto);
+        Post post = new Post(postRequestDto, user);
 
         //게시글 저장
         Post savePost = postRepository.save(post);
+
         return new PostResponseDto(savePost);
 
     }
@@ -47,41 +50,44 @@ public class PostService {
     }
 
     //게시물 선택 조회
-    public PostResponseDto getPost(Long id) {
+    public PostResponseDto getPost(Long id, UserDetailsImpl userDetails) {
         Post post = postRepository.findById(id).
                 orElseThrow(() -> new IllegalArgumentException("해당 회원이 존재하지 않습니다."));
 
-        return new PostResponseDto(post);
+        if (post.getWriter().equals(userDetails.getUsername())) {
+            return new PostResponseDto(post);
+        } else return new PostResponseDto("접근 할 수 없습니다. ");
     }
 
 
     //게시물 수정
     @Transactional
-    public Post updatePost(Long id, PostRequestDto postRequestDto) {
+    public PostResponseDto updatePost(Long id, PostRequestDto postRequestDto, UserDetailsImpl userDetails) {
+        //id 값 비교
         Post post = postRepository.findById(id).
                 orElseThrow(() -> new IllegalArgumentException("해당 회원이 존재하지 않습니다."));
 
-        post.checkPassword(postRequestDto.getPassword());
+        //게시글 이름과 회원 이름 비교
+       if (!post.getWriter().equals(userDetails.getUsername())) {
+           throw new RejectedExecutionException();
+       }
+           post.update(postRequestDto.getTitle(),
+                   postRequestDto.getContent());
 
-        post.update(postRequestDto.getTitle(),
-                postRequestDto.getWriter(),
-                postRequestDto.getContent(),
-                postRequestDto.getPassword());
-
-        return post;
+       return new PostResponseDto(post);
 
     }
 
     //게시물 삭제
-    public void deletePost(Long id, String password) {
+    public void deletePost(Long id, UserDetailsImpl userDetails) {
         Post post = postRepository.findById(id).
                 orElseThrow(() -> new IllegalArgumentException("해당 회원이 존재하지 않습니다."));
 
-        post.checkPassword(password);
+        if (post.getWriter().equals(userDetails.getUsername())) {
+            postRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("본인이 아니거나,관리자가 아니면 불가능.");
 
-
-
-        postRepository.deleteById(id);
-
+        }
     }
 }
